% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{aaties32}
\alias{aaties32}
\title{32 sets of randomly created alter-alter ties belonging to ego-centered networks}
\format{A data frame with 32 sets of alter-alter relations and 4 variables:
\describe{
\item{.EGOID}{ego identifier}
\item{.SRCID}{source alter ID}
\item{.TGTID}{target alter ID}
\item{weight}{weight of relation}
}}
\usage{
aaties32
}
\description{
32 sets of randomly created alter-alter ties belonging to ego-centered networks
}
\keyword{datasets}
