% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activation.R
\name{activate.egor}
\alias{activate.egor}
\title{Activate ego, alter or alter-alter tie data llevel of an egor .dataect}
\usage{
\method{activate}{egor}(.data, what)
}
\arguments{
\item{.data}{The \code{egor} .dataect.}

\item{what}{\code{Character} naming the level to activate, this can be "ego", "alter" or "aatie".}
}
\description{
This function activates one of the data levels of an egor .dataect, so that the dplyr verbs know which level to execute on.
}
\keyword{ego-centered}
\keyword{network}
