% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{add_ego_vars_to_long_df}
\alias{add_ego_vars_to_long_df}
\title{add_ego_vars_to_long_df}
\usage{
add_ego_vars_to_long_df(alteri.list, egos.df, ego.vars, netsize)
}
\arguments{
\item{alteri.list}{\code{List} of \code{data.frame}s containg alteri attributes.}

\item{ego.vars}{A \code{data.frame} of alteri attributes in the wide format.}

\item{netsize}{\code{Vector} containing values of network size per ego.}

\item{egos}{\code{Data frame} containg ego data (egos as cases)}
}
\description{
This function adds ego attributes to a 'alteri.list' object of ego-centered
networks. This is helpful if (multi-level) regressions are to be executed.
}
\keyword{internal}

