% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allbus_2010_simulated}
\alias{allbus_2010_simulated}
\title{Simulated Allbus 2010 Data}
\format{
A tibble/ data.frame of 100 simulated respondents/ rows and 981
variables/ columns. Each variable is a \verb{labelled dbl}.
}
\usage{
allbus_2010_simulated
}
\description{
A dataset simulated based on the the original Allbus
2010 SPSS data. The dataset simulates 100 respondents and does not resemble
any actual Allbus respondents. Each variable is randomly generated based on
the range of the original variables, co-varianaces between variables are
disregarded. The data’s purpose is purely to demonstrate how to technically
work with the Allbus data using egor and R - no analytical assumptions should
be made based on this data!
}
\details{
The dataset contains (simulated!) answers to two ego-centered name generators.
}
\keyword{datasets}
