% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alter.design.R
\docType{methods}
\name{alter_design}
\alias{alter_design}
\alias{alter_design.egor}
\alias{alter_design<-}
\alias{alter_design<-.egor}
\title{Set and query the alter nomination design}
\usage{
alter_design(x, ...)

\method{alter_design}{egor}(x, which, ...)

alter_design(x, ...) <- value

\method{alter_design}{egor}(x, which, ...) <- value
}
\arguments{
\item{x}{an \code{\link{egor}} object.}

\item{...}{arguments to be passed to methods}

\item{which}{name of the alter design setting to query or replace}

\item{value}{if \code{which} is specified, the new value of the
attribute; if not, a named list of settings that replace their
old values.}
}
\description{
Extract, set, or update the alter nomination design associated with
an ego-centered dataset.
}
