% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alters32}
\alias{alters32}
\title{32 sets of randomly created alters belonging to ego-centered networks}
\format{A data frame with 32 sets of up to 32 alters per egoID and 7 variables:
\describe{
\item{egoID}{ego identifier}
\item{ego.sex}{ego's gender}
\item{age}{ego's age}
\item{netsize}{size of ego's network}
\item{alterID}{alter identifier}
\item{alter.sex}{alter's gender}
\item{alter.age}{alter's age}
}}
\usage{
alters32
}
\description{
32 sets of randomly created alters belonging to ego-centered networks
}
\keyword{datasets}
