% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as_alters_df}
\alias{as_alters_df}
\alias{as_aaties_df}
\title{Create global alters and alter-alter relations dataframes from an \code{egor} object}
\usage{
as_alters_df(object, include.ego.vars = FALSE)

as_aaties_df(object, include.ego.vars = FALSE, include.alter.vars = FALSE)
}
\arguments{
\item{object}{An \code{egor} object.
a new variable with the specified name is created.}

\item{include.ego.vars}{Logical, specifying if ego variables should be included in the result.}

\item{include.alter.vars}{Logical, specifying if alter variables should be included in the result.}
}
\value{
A \code{tibble}.
}
\description{
Provided an egor-object, these functions create a 'global' \code{data.frame},
containing alter attributes, or alter-alter relations. The resulting dataframes
are useful for advanced analysis procedures, e.g. multi-level regressions.
}
\details{
These functions are convenience functions for egor's \code{as_tibble} method.
}
\examples{
# Load example data
data(egor32)

# Create global alters dataframes
as_alters_df(egor32)

# Create global alter-alter relaions dataframes
as_aaties_df(egor32)

# ... adding alter variables
as_aaties_df(egor32, include.alter.vars = TRUE)
}
