% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as_igraph}
\alias{as_igraph}
\alias{as_igraph.nested_egor}
\alias{as.igraph.egor}
\title{Convert \code{egor} object to \code{network} or \code{igraph} objects}
\usage{
as_igraph(
  x,
  directed = FALSE,
  include.ego = FALSE,
  ego.attrs = NULL,
  ego.alter.weights = NULL
)

\method{as_igraph}{nested_egor}(
  x,
  directed = FALSE,
  include.ego = FALSE,
  ego.attrs = NULL,
  ego.alter.weights = NULL
)

as.igraph.egor(
  x,
  directed = FALSE,
  include.ego = FALSE,
  ego.attrs = NULL,
  ego.alter.weights = NULL
)
}
\arguments{
\item{x}{An \code{egor} object.}

\item{directed}{Logical, indicating if alter-alter relations are directed.}

\item{include.ego}{\code{Logical.} Should ego be included?}

\item{ego.attrs}{Vector of names (character) or indices (numeric) of ego
variables that should be carried over to  the network/
igraph objects.}

\item{ego.alter.weights}{Vector of names (character) or indices (numeric) of
alter variables that should be carried over to the the
network/ igraph objects, as edge attributes of the ego-alter relations.}
}
\description{
These functions convert an \code{egor} object into a list of \code{network} or \code{igraph} objects.
By default ego itself is not included in the created objects, there is
a parameter (\strong{include.egor}) that allows for including ego.
}
\details{
The names of the variables specified in ego.attr and ego.alter.attr
need to be the same as the names of corresponding alter attributes,
in order for those variables to be complete in the resulting network/ igraph
object (see example).
}
