% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as_network}
\alias{as_network}
\alias{as.network.egor}
\title{Creates a list of statnet's network objects, from an
\code{egor} object.}
\usage{
as_network(x, directed = FALSE, include.ego = FALSE,
  ego.attrs = NULL, ego.alter.weights = NULL)

\method{as.network}{egor}(x, directed = FALSE, include.ego = FALSE,
  ego.attrs = NULL, ego.alter.weights = NULL)
}
\arguments{
\item{x}{An egor Object.}

\item{directed}{Logical.}

\item{include.ego}{Logical.}

\item{ego.attrs}{Names of ego variables.}

\item{ego.alter.weights}{Name of ego alter weight variable.}
}
\description{
Creates a list of statnet's network objects, from an
\code{egor} object.
}
