% Generated by roxygen2 (4.0.2): do not edit by hand
\name{broad.to.long}
\alias{broad.to.long}
\title{Transform 'broad' alter-level data to the 'long'-format}
\usage{
broad.to.long(items.df, netID, max.alteri, start.col, end.col)
}
\arguments{
\item{items.df}{A broad-format dataframe of ego-centric network data.}

\item{netID}{Variable containing netork IDs. (#!# Harmonise netID/ netID!!)}

\item{max.alteri}{A numeric for the maximum number of alteri.}

\item{start.col}{Number of first colum containg alter-alter relation data. #!# Should: Defaults to first column of \code{items.df}.}

\item{last.col}{Number of first colum containg alter-alter relation data. #!# Should: Defaults to last column of \code{items.df}.}
}
\description{
A function to transform a broad-format dataframe of ego-centric network data into a long-format data-frame, where every row represents one alter/dyad. In the created dataframe numerous networks can be distinguished by a network ID (netID).
}

