% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.three.files.R
\name{check.ID.integrity}
\alias{check.ID.integrity}
\title{Check for ID integrity}
\usage{
check.ID.integrity(egos, alters.df, edges, egoID = "egoID",
  alterID = "alterID")
}
\description{
Check for ID integrity
}
\keyword{internal}
