% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{col_idx}
\alias{col_idx}
\title{Obtain the index of a column in a data frame (or a list), producing
an error if there is a problem.}
\usage{
col_idx(name, df)
}
\arguments{
\item{name}{a character vector giving the names of the columns to look up.}

\item{df}{a \code{\link{data.frame}} or a \code{\link{list}} object.}
}
\value{
An integer giving the column index of the named column.
}
\description{
Obtain the index of a column in a data frame (or a list), producing
an error if there is a problem.
}
\note{
Numeric inputs for \code{name} are passed through, so this
function is safe to use if the input is already a column index.
}
\keyword{internal}
