% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{common_prefix}
\alias{common_prefix}
\title{Longest common prefix of a set of strings.}
\usage{
common_prefix(x)
}
\arguments{
\item{x}{a character vector.}
}
\value{
A character vector that is the longest common substring at
the start of each of the input vectors.
}
\description{
Longest common prefix of a set of strings.
}
\keyword{internal}
