% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comp.cat.counts.na}
\alias{comp.cat.counts.na}
\title{Set compositional results to NA if netsize is missing (NA), zero or not a number (NaN).}
\usage{
comp.cat.counts.na(cat.counts, netsize)
}
\arguments{
\item{cat.counts}{Results of the function comp.cat.counts}

\item{netsize}{Name of a variable in \code{broad} consisting of numerics for the network size of each network.}
}
\description{
This function sets compositional results to NA if netsize is missing (NA), zero or not a number (NaN).
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

