% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comp.cat.counts}
\alias{comp.cat.counts}
\title{Count category frequencies of an alter attribute for ego-centric-network data.}
\usage{
comp.cat.counts(long.df, var, egoID = "egoID", fun = fun.count)
}
\arguments{
\item{long.df}{A 'long' dataframe with alteri/dyads in rows.}

\item{egoID}{Name of the network-/ ego-ID variable in \code{broad}.}

\item{netsize}{Name of a variable in \code{broad} consisting of numerics for the network size of each network.}
}
\description{
This function counts the category frequencies of a variable representing alter attributes in ego-centric-network data.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

