% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comp.diversity}
\alias{comp.diversity}
\title{Calculate the network diversity of ego-centric networks for a given alter-attribute.}
\usage{
comp.diversity(cat.counts, netsize)
}
\arguments{
\item{cat.counts}{A category-count-dataframe usually generated with gen.egonet.agg().}

\item{netsize}{A vector of network sizes.}
}
\description{
This function calculates the network diversity of ego-centric networks for a given alter-attribute.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

