% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comp.ei}
\alias{comp.ei}
\title{Calculate the EI-Index of ego-centric networks.}
\usage{
comp.ei(cat.counts, v_ego, netsize)
}
\arguments{
\item{x}{A factor representing an alteri-attribute variable.}
}
\description{
This function calculates the EI-Index of ego-centric networks using a category-count-dataframe usually generated with gen.egonet.agg().
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

