% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{comp.homophily}
\alias{comp.homophily}
\title{Calculate the homophily of ego-centered networks.}
\usage{
comp.homophily(cat.counts, v_ego, netsize)
}
\arguments{
\item{cat.counts}{Results of the function comp.cat.counts}

\item{v_ego}{variable of an ego attribute corresponding to the}

\item{netsize}{Name of a variable in \code{broad} consisting of numerics for the network size of each network.}
}
\value{
EI Index values per ego/ network in a \code{dataframe}.
}
\description{
This function calculates the EI-Index of ego-centered networks using a "category-count-dataframe" usually generated with comp.cat.counts(). The EI-Index ranges from -1 to 1, where a positive value represents a heterogenous personal network in respect to a specified attribute of ego an its alters.
}
\keyword{internal}
