% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{comp_ei}
\alias{comp_ei}
\title{Calculate the EI-Indices of an \code{egor} object as a measurement of ego-alter homophily}
\usage{
comp_ei(object, alt.attr, ego.attr)
}
\arguments{
\item{object}{An \code{egor} object.}

\item{alt.attr}{A \code{character} naming the variable containing the alter-attribute.}

\item{ego.attr}{A \code{character} naming an ego attribute.}
}
\value{
A \code{numeric} vector.
}
\description{
\code{comp_ei()} calculates the EI-Index values as a measurement for ego-alter homophily.
}
\examples{
data("egor32")
comp_ei(egor32, "age", "age")
}
\keyword{analysis}
\keyword{ego-centered}
\keyword{network}
