% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composition.R
\name{composition}
\alias{composition}
\title{All-in-one ego-centric network composition function.}
\usage{
composition(long.df, v_alt, netsize, egoID = "egoID", v_ego = NULL,
  mode = "regular")
}
\arguments{
\item{long.df}{A 'long' dataframe with alteri in rows.}

\item{v_alt}{A character naming the variabe containg the alter-attribute.}

\item{netsize}{A vector of network sizes.}

\item{v_ego}{Character vector containing the ego attribute. Only needed if homophily should be calculated. Caution: Levels of v_alt and v_ego need to correspond.}

\item{mode}{A character. "regular" for a basic output, "all" for a complete output.}
}
\value{
Returns a dataframe with category counts, diversity and EI-Index values.
}
\description{
This function outputs a dataframe containg serveral compositional measures for ego-centric-network data.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

