% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{convert_egor}
\alias{convert_egor}
\alias{as_igraph}
\alias{as_network}
\title{Convert egor object to network or igraph objects}
\usage{
as_igraph(object, directed = FALSE, include.ego = FALSE, ego.attrs = NULL,
  ego.alter.weights = NULL)

as_network(object, directed = FALSE, include.ego = FALSE,
  ego.attrs = NULL, ego.alter.weights = NULL)
}
\arguments{
\item{object}{An egor object.}

\item{include.ego}{\code{Logical.} Should ego be included?}

\item{ego.attr}{Vector of names (character) or indices (numeric) of ego
variables that should be carried over to  the network/
igraph objects.}

\item{ego.alter.attr}{Vector of names (character) or indices (numeric) of
alter variables that should be carried over to the the
network/ igraph objects, as edge attributes of the ego-alter relations.}
}
\description{
These functions convert an egor object into a list of network or igraph objects.
By default ego itself is not included in the created objects, but there is
a parameter (\strong{include.egor}) that allows for including ego.
}
\details{
The names of the variables specified in ego.attr and ego.alter.attr
need to be the same as the names of corresponding alter attributes,
in order for those variables to be complete in the resulting network/ igraph
object (see example).
}
\section{Functions}{
\itemize{
\item \code{as_igraph}: Creates a list of igraph objects from an
egor object.

\item \code{as_network}: Creates a list of statnet's network objects, from an
egor object.
}}

