% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/helper.R
\name{dyad.poss}
\alias{dyad.poss}
\alias{poss.dyads.wide}
\title{Helper functions for ego centric network analysis}
\usage{
dyad.poss(max.alteri, directed = F)

poss.dyads.wide(max.alteri)
}
\arguments{
\item{max.alteri}{A \code{numeric} giving the maximum number of alteri.}

\item{directed}{A \code{logical} value indicating directedness of alter-alter
data.}
}
\value{
\code{dyad.poss} returns the count of possible edges.
}
\description{
Helper functions for ego centric network analysis
}
\section{Functions}{
\itemize{
\item \code{poss.dyads.wide}: Generates a \code{data.frame} marking possible dyads in
a wide alter-alter relation \code{data.frame}. Row names corresponds to the
network size. This is useful for sanitizing alter-alter relations in the wide
format.
}}

