% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_egor.R
\name{edgelist_to_wide}
\alias{edgelist_to_wide}
\title{Transforms edge lists to alter-alter wide format data.}
\usage{
edgelist_to_wide(edges, src_tgt_id = c(".SRCID", ".TGTID"), weight = "weight")
}
\arguments{
\item{edges}{List of \code{data.frames} containing edge lists.}
}
\description{
Transforms edge lists to alter-alter wide format data.
}
\keyword{ego-centered}
\keyword{internal}
\keyword{network}
