% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edges.attributes.to.network}
\alias{edges.attributes.to.network}
\title{edges.attributes.to.network}
\usage{
edges.attributes.to.network(elist, attributes)
}
\arguments{
\item{elist}{[dataframe]}

\item{attributes}{[dataframe]}
}
\description{
This function generates one igraph object from an edgelist and a dataframe
alteri attributes.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

