% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.random.egonet.data.R
\name{edges.to.wide}
\alias{edges.to.wide}
\title{Transfroms edge lists to alter-alter wide format data.}
\usage{
edges.to.wide(edges)
}
\arguments{
\item{edges}{List of \code{data.frames} containg edge lists.}
}
\description{
Only works properly, if the netsize of all networks is constant.
}
\keyword{ego-centric}
\keyword{network}
\keyword{sna}

