% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edges32}
\alias{edges32}
\title{32 sets of randomly created alter-alter ties belonging to ego-centered networks}
\format{A data frame with 32 sets of alter-alter relations and 3 variables:
\describe{
\item{egoID}{ego identifier}
\item{Source}{source alter ID}
\item{Target}{target alter ID}
\item{weight}{weight of relation}
}}
\usage{
edges32
}
\description{
32 sets of randomly created alter-alter ties belonging to ego-centered networks
}
\keyword{datasets}
