% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego.design.R
\docType{methods}
\name{ego.design}
\alias{ego.design}
\alias{ego.design.egor}
\alias{ego.design<-}
\alias{ego.design<-.egor}
\title{Set and query the ego sampling design}
\usage{
ego.design(x, ...)

\method{ego.design}{egor}(x, ...)

ego.design(x, ...) <- value

\method{ego.design}{egor}(x) <- value
}
\arguments{
\item{x}{an \code{\link{egor}} object.}

\item{value}{either \code{survey.design} object (like one constructed by
\code{\link[=svydesign]{svydesign()}}) or a \code{\link{list}} of arguments to \code{\link[=svydesign]{svydesign()}}
specifying the sampling design for the egos. If the arguments are
formulas, they can refer to columns (ego attributes) of \code{x}.}
}
\description{
Extract, set, or update the \code{\link{svydesign}} associated with an
egocentric dataset.
}
\note{
This can be useful for adjusting or reinitializing the ego
design information after the underlying ego attributes had been
modified.
}
