% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{ego_density}
\alias{ego_density}
\title{Calculate the density for n+many ego-centered networks.}
\usage{
ego_density(x, alters = NULL, egoID = "egoID", weight = NULL,
  max.netsize = NULL, directed = FALSE)
}
\arguments{
\item{x}{Either an \code{egor} object or a \code{data.frame/ list} containg alter-alter
relations.}

\item{alters}{\code{data.frame/ list} containg alter-alter relation data 
(ignore if x is an \code{egor} object).}

\item{egoID}{Name of ego ID variable. Only needs to be specified if alter
and alter-alter data is provided in global data.frames.}

\item{weight}{\code{Character} naming a variable containing the weight values
of relations. Weights should range from 0 to 1.}

\item{max.netsize}{Optional parameter. Constant value used if the
number of alters whose relations were collected is limited.}

\item{directed}{logical indicating if the alter-alter relation data/ edges 
are directed or undirected.}
}
\value{
returns a \code{vector} of network density values.
}
\description{
This function uses an \code{egor} object and calculates the density of all 
the ego-centered networks listed in the 'egor' object. Instead of an 
\code{egor} object, alter and alter-alter data can be provided as \code{lists}
or \code{data.frames}.
}
\examples{
data("egor32")
egor_density(egor32)
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}
