% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{egoR.avg.rel}
\alias{egoR.avg.rel}
\title{Average Alteri Relation Count}
\usage{
egoR.avg.rel(egoR, netsize = NULL)
}
\arguments{
\item{netsize}{\code{vector} containing values of network size per ego.}

\item{A}{\code{list} of six objects - the **egoR** object: (1) egos.df: \code{dataframe} of all
egos and their attributes; (2) alteri.df: \code{dataframe} of all alteri;
(3) alteri.list: \code{list} of \code{dataframes} of all alteri per ego;
(4) edges: \code{list} of \code{dataframes} with edge lists per network; (5)
graphs: \code{list} of igraph objects; (6) results: a result \code{dataframe},
pre-populated with the network size of each network}
}
\description{
Average Alteri Relation Count
}
\seealso{
\code{\link{egoR.density}}
}

