% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/density.R
\name{egoR.density}
\alias{egoR.density}
\title{Calculate the density for several ego-centered networks.}
\usage{
egoR.density(egoR, weight = "weight", max.netsize = NULL)
}
\arguments{
\item{egoR}{A list of six data objects, created by one of egonetR's import
functions.}

\item{max.netsize}{Optional parameter. Only needed if the density information
was only collected for a subsample of all alteri per network.}
}
\value{
returns a \code{vector} of network density values.
}
\description{
This function uses an 'egoR' object (list of data objects created by an
egonetR import function) and calculates the density of all the ego-centered
networks listed in the 'egoR' object.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

