% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizz.R
\name{egoR.vis.wizzard}
\alias{egoR.vis.wizzard}
\title{Network Visualisation Wizzard (egoR)}
\usage{
egoR.vis.wizzard(egoR)
}
\arguments{
\item{egoR}{A list of six data objects, created by one of egonetR's import 
functions.}
}
\value{
Opens a shiny app.
}
\description{
This function uses an 'egoR' object (list of data objects created by an 
egonetR import function), extracts the graphs object and uses it to visualise
all networks contained (Shiny App).
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

