% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{egor_options}
\alias{egor_options}
\title{Display names and values of global egor options.}
\usage{
egor_options()
}
\description{
Display names and values of global egor options.
}
\details{
Currently egor makes use of the following global options. Use
\code{options()} to change values.
\subsection{egor.print.rows.active.level}{
\itemize{
\item \code{Numeric}. Amount of rows to display when printing the active level of an \code{egor} object.
}
}

\subsection{egor.print.rows.inactive.level}{
\itemize{
\item \code{Numeric}. Amount of rows to display when printing the inactive levels of an \code{egor} object.
}
}

\subsection{egor.print.switch.active.level.to.top}{
\itemize{
\item \code{Logical}. When printing an egor object, should the active data-level always
be printed first?
}
}

\subsection{egor.return.results.with.design}{
\itemize{
\item \code{Logical}. \code{egor} functions that return ego-level results (e.g. one value per
ego) return a \code{svy_tbl} object containing the \code{ego_design()}, when this is set to \code{TRUE}.
}
}
}
