% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizz.R
\name{egor_vis_wizzard}
\alias{egor_vis_wizzard}
\title{Network Visualisation Wizzard (egor)}
\usage{
egor_vis_wizzard(object)
}
\arguments{
\item{object}{An egor object.}
}
\value{
Opens an interactive Browser Application.
}
\description{
This function uses an 'egor' object (list of data objects created by an
egor import function), extracts the graphs object and uses it to visualise
all networks contained in an interactive Browser Application (R-Shiny).
}
\examples{
\dontrun{
data("egor32")
egor.vis.wizzard(egor32)
}
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}
