% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egos32}
\alias{egos32}
\title{32 randomly created egos belonging to ego-centered networks}
\format{A data frame with 32 sets of alter-alter relations and 3 variables:
\describe{
\item{egoID}{ego identifier}
\item{sex}{ego's gender}
\item{age}{ego's age}
\item{netsize}{size of ego's network}
}}
\usage{
egos32
}
\description{
32 randomly created egos belonging to ego-centered networks
}
\keyword{datasets}
