% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.three.files.R
\name{ensure.ID.integrity}
\alias{ensure.ID.integrity}
\title{Ensure ID integrity}
\usage{
ensure.ID.integrity(egos, alteri.df, edges, egoID = "egoID",
  alterID = "alterID")
}
\description{
Ensure ID integrity
}
\keyword{internal}

