% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ei.R
\name{fragmentations}
\alias{fragmentations}
\title{Fragmentations of a list of ego-centric networks}
\usage{
fragmentations(edges_)
}
\arguments{
\item{edges_}{\code{List} of edgelist-\code{dataframes}}
}
\description{
A longer description will appear here in the future.
}
\keyword{ego-centric}
\keyword{network}
\keyword{sna}

