% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composition.R
\name{fun.prop}
\alias{fun.prop}
\title{Generate proportional variables for each category of a variable. Used by comp.cat.counts().}
\usage{
fun.prop(x)
}
\arguments{
\item{x}{A factor representing an alteri-attribute variable.}
}
\description{
This function generates Variables of the proportional frequency of each category of a variable. To use on data aggregated from the long ego-centered network data format. #!# prop.table() produces errors here, when netsize and total of frequencies per variable do not align.
#'
}
\keyword{analysis}
\keyword{ego-centered}
\keyword{network}

