% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.random.egonet.data.R
\name{generate.sample.edge.list}
\alias{generate.sample.edge.list}
\title{Generate a random edge list for one network.}
\usage{
generate.sample.edge.list(max.alteri)
}
\arguments{
\item{max.alteri}{\code{Numeric} indicating maximum number of alteri.}
}
\description{
Here should be a a longer description of this function.
}
\keyword{ego-centric}
\keyword{internal}
\keyword{network}

