% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/create.random.egonet.data.R
\name{generate.sample.ego.data}
\alias{generate.sample.ego.data}
\title{Generate random ego-centric-network data.}
\usage{
generate.sample.ego.data(net.count, h, netsize = NULL)
}
\arguments{
\item{net.count}{Number of networks/ egos to generate.}

\item{max.alteri}{Maximum size of networks.}
}
\description{
This function generates random ego-centric-network data for a specified number of networks with a maximum network size. The network size of the generated networks is a normal distribution with sd=5.
}
\keyword{ego-centric}
\keyword{network}
\keyword{sna}

