% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{helper}
\alias{create.edge.names.wide}
\alias{din.page.dist}
\alias{dyad.poss}
\alias{dyads.possible.between.groups}
\alias{helper}
\alias{sanitize.wide.edges}
\title{General helper functions}
\usage{
dyad.poss(max.alteri, directed = F)

sanitize.wide.edges(max.alteri)

create.edge.names.wide(x)

dyads.possible.between.groups(x, y)

din.page.dist(x)
}
\arguments{
\item{max.alteri}{A \code{numeric} giving the maximum number of alteri.}

\item{directed}{A \code{logical} value indicating directedness of alter-alter
data.}

\item{x}{\code{Numeric}.}

\item{y}{\code{Numeric}.}
}
\description{
Helper functions for ego centric network analysis
}
\section{Functions}{
\itemize{
\item \code{dyad.poss}: Returns the count of possible edges in an
undirected or directed, ego-centered network, based on the number of alteri.

\item \code{sanitize.wide.edges}: Generates a \code{data.frame} marking possible dyads in 
a wide alter-alter relation \code{data.frame}. Row names corresponds to the 
network size. This is useful for sanitizing alter-alter relations in the wide 
format.

\item \code{create.edge.names.wide}: Creates a \code{vector} of names for variables 
containing data on alter-alter relations/ dyads in ego-centered networks.

\item \code{dyads.possible.between.groups}: Calculates the possible edges between members of 
different groups in an ego-centered network.

\item \code{din.page.dist}: Calculates the optimal distribution of a number of 
equally sized objects on a DIN-Norm DIN 476 (i.e. DIN A4) page in landscape 
view.
}}

