% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{homophily_ei}
\alias{homophily_ei}
\title{Calculate the EI-Indices of an egor object as a measurement of ego-alter homophily}
\usage{
homophily_ei(onject, alt.attr, ego.attr)
}
\arguments{
\item{alt.attr}{A \code{character} naming the variable containg the alter-attribute.}

\item{ego.attr}{A \code{character} naming an ego attribute.}

\item{object}{An egor object.}
}
\value{
A \code{numeric vector}.
}
\description{
\code{homophily_ei()} calculates the EI-Index values as a measurement for ego-alter homophily.
}
\examples{
data("egor32")
homophily_ei(egor32, "age", "age")
}
\keyword{analysis}
\keyword{ego-centered}
\keyword{network}
