% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ego_gram.R
\name{layout_egogram}
\alias{layout_egogram}
\title{Create layout for an egogram}
\usage{
layout_egogram(altID, venn_var, pie_var)
}
\arguments{
\item{altID}{Vector of alter IDs.}

\item{venn_var}{Vector of values representing alter groups corresponding with
venns in an egogram.}

\item{pie_var}{Vector of values representing alter groups corresponding with
pieces of pie in an egogram.}
}
\value{
A dataframe with three columns: x, y and altID.
}
\description{
This creates pairs of x and y coordinates for a egogram, accompanied by
alter IDs for each coordinate pair.
}
