% Generated by roxygen2 (4.0.2): do not edit by hand
\name{long.df.to.list}
\alias{long.df.to.list}
\title{Trim/ listify ego-centric network data}
\usage{
long.df.to.list(long, wide, netsize, egoID, back.to.df = F)
}
\arguments{
\item{long}{A 'long' dataframe with alteri/dyads in rows.}

\item{wide}{A 'wide' dataframe with networks in rows.}

\item{netsize}{\code{vector} containing values of network size per ego.}

\item{egoID}{Name of the variable identifying egos (default: "egoID").}

\item{back.to.df}{If \code{TRUE} a dataframe is returned, if \code{FALSE} a
list. Defaults to \code{FALSE}.}
}
\value{
Returns a \code{list} of \code{dataframes} where every
\code{dataframe} represents one ego/ network and the rows in the
\code{dataframe} represents one alter. If the \code{back.to.df} parameter is
called the \code{list} entries are combined to one \code{dataframe}, in the
'long' format.
line represents
}
\description{
This function deletes empty alteri rows and it can be used to transform the
dataframe into a list, with entries for each network.
}

