% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.random.egonet.data.R
\name{make_edge_list}
\alias{make_edge_list}
\title{Generate a random edge list for one network.}
\usage{
make_edge_list(netsize)
}
\arguments{
\item{netsize}{\code{Numeric} indicating maximum number of alters.}
}
\description{
Generate a random edge list for one network.
}
\keyword{ego-centered}
\keyword{internal}
\keyword{network}
