% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.random.egonet.data.R
\name{make_egor}
\alias{make_egor}
\title{Generate random ego-centered-network data.}
\usage{
make_egor(net.count, max.alters, netsize = NULL, plot = FALSE)
}
\arguments{
\item{net.count}{Number of networks/ egos to generate.}

\item{max.alters}{Maximum size of networks.}

\item{netsize}{\code{Numeric} for fixed network sizes.}

\item{plot}{whether to plot the network size distribution.}
}
\description{
This function generates random ego-centered-network data for a specified number of networks with a maximum network size. The network size of the generated networks is a normal distribution with sd=5.
}
\keyword{ego-centered}
\keyword{network}
\keyword{random}
