% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.three.files.R
\name{order.edge.list.columns}
\alias{order.edge.list.columns}
\title{Order edge list colums with source and target at the beginning.}
\usage{
order.edge.list.columns(edges, source_, target)
}
\description{
Order edge list colums with source and target at the beginning.
}
\keyword{internal}
