% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.egonet.one.file}
\alias{read.egonet.one.file}
\title{Import ego-centric network data from one file ("broad"-format)}
\usage{
read.egonet.one.file(egos, netsize, netID = "netID", attr.start.col,
  attr.end.col, dy.max.alteri, dy.first.var)
}
\arguments{
\item{egos}{}

\item{netsize}{}

\item{netID}{}

\item{attr.start.col}{}

\item{dy.max.alteri}{}

\item{dy.first.var}{}
}
\description{
This function imports ego-centric network data, that is supplied as one file,
also known as the "broad"-foramt.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{netowrk}

