% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.egonet}
\alias{read.egonet}
\title{Import raw ego-centric network data.}
\usage{
read.egonet(alteri, y)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to import raw ego-centric network data. See X for supported formats.
}
\examples{
cat_function()
}
\keyword{cats}

