% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.egonet.two.files}
\alias{read.egonet.two.files}
\title{Import ego-centric network data from one file ("broad"-format)}
\usage{
read.egonet.two.files(egos, long, netsize = NULL, netID = "netID",
  alterID = NULL, dy.max.alteri, dy.first.var)
}
\arguments{
\item{egos}{}

\item{netsize}{}

\item{netID}{}

\item{dy.max.alteri}{}

\item{dy.first.var}{}
}
\description{
This function imports ego-centric network data, stored in two files, where
one file contains the ego attributes, items and the dyads and the other file
contains the alteri attributes. This format is also known as the
"long"-format.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{netowrk}

