% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{return_results}
\alias{return_results}
\title{Returns results inheriting \code{srvyr} design if the input egor object has a an
\code{ego_design} and global option "egor.return.results.with.design" is \code{TRUE} or
'NULL'.}
\usage{
return_results(x, results)
}
\arguments{
\item{x}{Original \code{egor} object, as submitted in call to parent function.}

\item{results}{\code{data.frame} with \code{.egoID} column and a column that hold the
ego-level results.}
}
\description{
Returns results inheriting \code{srvyr} design if the input egor object has a an
\code{ego_design} and global option "egor.return.results.with.design" is \code{TRUE} or
'NULL'.
}
