% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate_to_equilibrium}
\alias{rotate_to_equilibrium}
\title{Rotate circle or star graph layout so that it 'stands' on two bottom nodes.}
\usage{
rotate_to_equilibrium(xy_mat)
}
\arguments{
\item{xy_may}{A two column matrix with x and y coordinates containing the graph layout.}
}
\description{
Rotate circle or star graph layout so that it 'stands' on two bottom nodes.
}
\keyword{internal}
