% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate_xy}
\alias{rotate_xy}
\title{Rotate a matrix of x and y coordinates by a radian}
\usage{
rotate_xy(xy_mat, rad, direction = 1)
}
\arguments{
\item{xy_mat}{A two column matrix with x and y coordinates.}

\item{rad}{\code{Numeric.} Rotation angle in radian.}

\item{direction}{\code{Numeric.} Rotation direction, \code{1} for clockwise \code{-1} for counter-clockwise.}
}
\description{
Rotate a matrix of x and y coordinates by a radian
}
\keyword{internal}
