% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{rowwise_egor}
\alias{rowwise_egor}
\title{Group input by rows (egor)}
\usage{
rowwise_egor(data, ...)
}
\arguments{
\item{data}{An \code{egor} object.}

\item{...}{Please refer to the
documentation of \code{\link[dplyr:rowwise]{dplyr::rowwise()}} for further explanations.}
}
\description{
This is a egor wrapper for \code{dplyr::rowwise()} as it is currently not possible
to define methods for it, as it is not a generic. Please refer to the
documentation of \code{\link[dplyr:rowwise]{dplyr::rowwise()}} for further explanations.
}
