% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.egor.R
\name{subset.egor}
\alias{subset.egor}
\alias{[.egor}
\title{Filter and Subset Egocentric Datasets}
\usage{
\method{subset}{egor}(x, subset, aspect = c("egos", "alters", "ties"), ...)

\method{[}{egor}(x, i, j, aspect = c("egos", "alters", "ties"), ...)
}
\arguments{
\item{x}{an \code{\link[=egor]{egor()}} object.}

\item{subset}{an expression evaluated on the columns of \code{\link[=egor]{egor()}}
(as in the eponymous argument of \code{\link[=subset]{subset()}}), specifying which
egos, alters, or alter-alter ties to keep; output format depends
on \code{aspect}: \describe{

\item{\code{"egos"}}{either an integer vector of indices specifying
which egos to select or a logical vector of length \code{nrow(x)}
specifying which rows should be kept.}

\item{\code{"alters"}}{\emph{for each ego's row}, either an integer vector of
indices specifying which alters to select or a logical vector of
length \code{nrow(x)} specifying which alters should be kept.}

\item{\code{"ties"}}{\emph{for each ego's row}, either an integer vector of
indices specifying which alter-alter ties to select or a logical
vector of length \code{nrow(x)} specifying which alter-alter ties should
be kept.}

}}

\item{aspect}{a selector of which aspect of an egocentric dataset
to affect: the egos, the alters or the (alter-alter) ties. Note
that only one can be affected at a time.}

\item{...}{extra arguments; currently unused.}

\item{i}{depends on \code{aspect}: \describe{

\item{\code{"egos"}}{either an integer vector of indices specifying
which egos to select or a logical vector of length \code{nrow(x)}
specifying which rows should be kept.}

\item{\code{"alters"}}{a ragged array (a \code{\link[=list]{list()}}) of length \code{nrow(x)},
either of integer vectors of indices specifying which alters to
select for the corresponding ego or of logical vectors of length
\code{nrow(x$.alters[k,,drop=FALSE])} specifying which alters should be kept.}

\item{\code{"ties"}}{a ragged array (a \code{\link[=list]{list()}}) of length \code{nrow(x)},
either of integer vectors of indices specifying which alter-alter
ties to select for the corresponding ego or of logical vectors of
length \code{nrow(x$.alter_ties[k,,drop=FALSE])} specifying which ties
should be kept}

}}

\item{j}{either an integer vector specifying which columns of the
filtered structure (egos, alters, or ties) to select, or a
logical vector specifying which columns to keep.}
}
\value{
An \code{\link[=egor]{egor()}} object.
}
\description{
Functions to index and take subsets of \code{\link[=egor]{egor()}} objects: manipulate
egos, alters, or alter-alter ties.
}
\details{
Removing or duplicating an ego will also remove or duplicate their
alters and ties.
}
