% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egor.R
\docType{methods}
\name{summary.egor}
\alias{summary.egor}
\alias{print.egor}
\title{Methods to print and summarize \code{\link{egor}} objects}
\usage{
\method{summary}{egor}(object, ...)

\method{print}{egor}(x, ...)
}
\arguments{
\item{object, x}{an \code{\link{egor}} object.}

\item{...}{additional arguments, either unused or passed to lower-level functions.}
}
\description{
Methods to print and summarize \code{\link{egor}} objects
}
