% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{trim_aaties}
\alias{trim_aaties}
\title{Trims alter-alter ties of alters that are missing/ deleted from alters data.}
\usage{
trim_aaties(object)
}
\arguments{
\item{object}{An \code{egor} object.}
}
\value{
An \code{egor} object with trimmed alter-alter ties (.aaties).
}
\description{
This is used in the background by \code{dplyr} methods, to maintain the alter-alter
ties according to changes made to the ego and alter data levels.
}
