% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.R
\name{trim_alters}
\alias{trim_alters}
\title{Trims alters that are missing/ deleted from ego data.}
\usage{
trim_alters(object)
}
\arguments{
\item{object}{An \code{egor} object.}
}
\value{
An \code{egor} object with trimmed alter-alter ties (.aaties).
}
\description{
This is used in the background by \code{dplyr} methods, to maintain the alter
ties according to changes made to the ego data level.
}
