% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.graphs.R
\name{vis.clustered_graphs}
\alias{vis.clustered_graphs}
\title{Visualise clustered graphs}
\usage{
vis.clustered_graphs(graphs, node.size.multiplier = 1, node.min.size = 0,
  node.max.size = 200, edge.width.multiplier = 30, center = 1,
  label.size = 0.8, labels = F, legend.node.size = 45, to.pdf = F)
}
\arguments{
\item{graphs}{\code{List} of \code{graph} objects, representing the clustered
graphs.}

\item{node.size.multiplier}{\code{Numeric} used to multiply the node diameter
of visualised nodes.}

\item{node.min.size}{\code{Numeric} indicating minimum size of plotted
nodes}

\item{node.max.size}{\code{Numeric} indicating maximum size of plotted
nodes}

\item{edge.width.multiplier}{\code{Numeric} used to mutliply the edge width.}

\item{center}{\code{Numeric} indicating the vertex to be plotted in center.}

\item{label.size}{\code{Numeric}.}

\item{labels}{\code{Boolean}. Plots with turned off labels will be preceeded
by a 'legend' plot giving the labels of the vertices.}

\item{legend.node.size}{\code{Numeric} used as node diameter of legend graph.}

\item{to.pdf}{\code{Boolean}.}
}
\value{
\code{vis.clustered_graphs} plots
a \code{list} of \code{igraph} objects created by the \code{clustered_graphs}
function.

\code{clustered_graphs} returns a list of graph objects representing
the clustered ego-centered network data;
}
\description{
\code{vis.clustered_graphs} visualises clustered_graphs using a list of
clustered graphs created with \code{\link{clustered_graphs}}
created clustered graph objects.
}
\examples{
data("egor32")

# Simplify networks to clustered graphs, stored as igraph objects
graphs <- clustered_graphs(egor32, "alter.age") 

# Visualise
vis.clustered_graphs(graphs)
}
\references{
Brandes, U., Lerner, J., Lubbers, M. J., McCarty, C., & Molina,
J. L. (2008). Visual Statistics for Collections of Clustered Graphs. 2008
IEEE Pacific Visualization Symposium, 47-54.
}
\seealso{
\code{\link{clustered_graphs}} for creating clustered graphs objects
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}
