% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizz.R
\name{vis.wizzard}
\alias{vis.wizzard}
\title{Network Visualisation Wizzard}
\usage{
vis.wizzard()
}
\value{
returns a \code{vector} of network density values.
}
\description{
This function uses an object named 'graphs' (list of data objects created by an 
egonetR import function), extracts the graphs object and uses it to visualise
all networks contained (Shiny App).
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

