% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{wide.dyads.to.edgelist.regex}
\alias{wide.dyads.to.edgelist.regex}
\title{Transform wide alter-alter data to an edge list.
A regex based implementation to convert a wide list to an edgelist.}
\usage{
wide.dyads.to.edgelist.regex(e.wide, aa.regex, netsize)
}
\arguments{
\item{e.wide}{A dataframe containing the alter-alter relation data
in the 'wide' format.}

\item{aa.regex}{A Perl regular expression with name capture,
intended to be run on column names and capturing via named
capture the following regex groups: \code{"attr"}, \code{"src"}, and
\code{"tgt"}, representing the edge attribute being captured, the
source (or the first alter identified), and the target (or the
second alter identified) of the edge, respectively. See \code{\link{regex}}
for more information.}

\item{netsize}{\code{Vector} containing values of network size per ego.}
}
\description{
Transform wide alter-alter data to an edge list.
A regex based implementation to convert a wide list to an edgelist.
}
\keyword{internal}
