% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{wide.to.long}
\alias{wide.to.long}
\title{Transform 'wide' alter-level data to the 'long'-format}
\usage{
wide.to.long(items.df, egoID = "egoID", max.alteri, start.col, end.col,
  ego.vars = NULL, var.wise = F)
}
\arguments{
\item{items.df}{A wide-format dataframe of ego-centric network data.}

\item{egoID}{Name of the variable identifying egos (default: "egoID").}

\item{max.alteri}{A numeric for the maximum number of alteri.}

\item{start.col}{Number of first colum containg alter-alter relation data.
#!# Should: Defaults to first column of \code{items.df}.}

\item{ego.vars}{Ego variables to transfer from the wide-format
\code{data.frame}, for i.e. multi-level regression analysis.}

\item{var.wise}{a logical value indicating wheter the alter attributes are
stored variable-wise, if FALSE alter-wise storage is assumed.}

\item{last.col}{Number of first colum containg alter-alter relation data.
#!# Should: Defaults to last column of \code{items.df}.}
}
\description{
A function to transform a wide-format dataframe of ego-centric network data
into a long-format data-frame, where every row represents one alter/dyad. In
the created dataframe numerous networks can be distinguished by a network ID
(egoID).
}

