% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{add_ego_vars_to_long_df}
\alias{add_ego_vars_to_long_df}
\title{add_ego_vars_to_long_df}
\usage{
add_ego_vars_to_long_df(alteri.list, egos.df, ego.vars, netsize)
}
\arguments{
\item{ego.vars}{}

\item{netsize}{\code{vector} containing values of network size per ego.}

\item{long.list}{}

\item{egos}{\code{dataframe} containg ego data (egos as cases)}
}
\description{
This function adds ego attributes to a 'long.list' object of ego-centered
networks. This is helpful if (multi-level) regressions are to be executed.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

