% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alters32}
\alias{alters32}
\title{32 sets of randomly created alters belonging to ego-centered networks}
\format{
A data frame with 32 sets of up to 32 alters per egoID and 7 variables:
\describe{
\item{.ALTID}{alter identifier}
\item{.EGOID}{ego identifier}
\item{age}{age in categories}
\item{age.years}{age in years}
\item{country}{country}
\item{income}{income}
\item{sex}{gender}
}
}
\usage{
alters32
}
\description{
32 sets of randomly created alters belonging to ego-centered networks
}
\keyword{datasets}
