% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{as_alts_df}
\alias{as_alts_df}
\alias{as_ties_df}
\title{Create global alters and alter-alter relations dataframes from an egor object}
\usage{
as_alts_df(object, egoID = "egoID", include.ego.vars = F)

as_ties_df(object, egoID = "egoID", include.ego.vars = F,
  include.alt.vars = F, aatie_vars = c(".srcID", "tgtID"))
}
\arguments{
\item{object}{An egor object.}

\item{egoID}{Character, naming the variable used to identify egos. If this variable does not exist,
a new variable with the specified name is created.}

\item{include.ego.vars}{Logical, specifying if ego variables should be included in the result.}

\item{include.alt.vars}{Logical, specifying if alter variables should be included in the result.}
}
\description{
Provided an egor-object, these functions create a 'global' \code{data.frame},
containing alter attributes, or alter-alter relations. The resulting dataframes
are useful for advanced analysis procedures, e.g. multi-level regressions.
}
\examples{
# Load example data
data(egor32)

# Create global alters dataframes
# ... creating a new egoID
as_alts_df(egor32)

# ... without creating a new egoID
as_alts_df(egor32, F)

# ... keeping ego variables
as_alts_df(egor32, F, T) 

# Create global alter-alter relaions dataframes
as_ties_df(egor32)

# ... adding alter variables
as_ties_df(egor32, include.alt.vars = T)
}
