% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.graphs.R
\name{clustered.graphs}
\alias{clustered.graphs}
\title{Cluster ego-centered networks by a grouping factor}
\usage{
clustered.graphs(alters.list, edges.list, clust.groups)
}
\arguments{
\item{alters.list}{\code{List} of \code{data frames} containing the alters 
data.}

\item{edges.list}{\code{List} of \code{data frames} containing the edge 
lists (= alter-alter relations).}

\item{clust.groups}{A \code{character} naming the \code{factor} variable defining the groups.}
}
\value{
\code{clustered.graphs} returns a list of graph objects representing 
the clustered ego-centered network data;
}
\description{
The idea of clustered graphs is to reduce the complexity of an ego-centered network
graph by visualising its group aggregated form. It is developed by
Lerner et al. (2008). It helps to discover and visualise structural and 
compostional properties of ego-centered networks, based on a pre-defined
factor variable on the alter level. clustered.graphs() calculates group sizes,
inter- and intragroup densities and these informations in a \code{list} of
\code{igraph} objects.
}
\examples{
data("egoR32")

# Simplify networks to clustered graphs, stored as igraph objects
graphs <- clustered.graphs(egoR32$alters.list, egoR32$edges, "alter.age") 

# Visualise
vis.clustered.graphs(graphs)
}
\references{
Brandes, U., Lerner, J., Lubbers, M. J., McCarty, C., & Molina, 
J. L. (2008). Visual Statistics for Collections of Clustered Graphs. 2008 
IEEE Pacific Visualization Symposium, 47-54.
}
\seealso{
\code{\link{vis.clustered.graphs}} for visualising clustered graphs
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}
