% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustered.graphs.R
\name{clustered.graphs}
\alias{clustered.graphs}
\alias{vis.clustered.graphs}
\title{Cluster ego-centered networks by a factor and visualise the results}
\usage{
clustered.graphs(alteri.list, edges.list, clust.groups)

vis.clustered.graphs(graphs, vertex.min.size, vertex.max.size, center = 1,
  labels = F, to.pdf = T)
}
\arguments{
\item{alteri.list}{\code{List} of \code{data frames} containing the alteri 
data.}

\item{edges.list}{\code{List} of \code{data frames} containing the edge 
lists.}

\item{clust.groups}{A \code{factor} variable the build the clusters from.}

\item{graphs}{\code{List} of \code{graph} objects, representing the clustered
graphs.}

\item{vertex.min.size}{\code{Numeric} indicating minimum size of plotted 
vertices.}

\item{vertex.max.size}{\code{Numeric} indicating maximum size of plotted 
vertices.}

\item{center}{\code{Numeric} indicating the vertex to be plotted in center.}

\item{labels}{\code{Boolean}. Plots with turned off labels will be preceeded 
by a 'legend' plot giving the labels of the vertices.}

\item{to.pdf}{\code{Boolean}.}
}
\value{
\code{clustered.graphs} returns a list of graph objects representing 
the clustered ego-centered network data; \code{vis.clustered.graphs} plots
the clustered graphs (into a pdf File if indicated)
}
\description{
The idea of clustered graphs in ego-centered network analysis is developed by
Brandes et al. (2008). It helps to discover and visualise structural and 
compostional properties of ego-centered networks, based on a pre-defined
factor variable on the alter level, i.e. the country of origin.
}
\section{Functions}{
\itemize{
\item \code{vis.clustered.graphs}: Visualises clustered.graphs using the formerly
created clustered graph objects.
}}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

