% Generated by roxygen2 (4.0.2): do not edit by hand
\name{comp.cat.counts}
\alias{comp.cat.counts}
\title{Count category frequencies of an alter attribute for ego-centric-network data.}
\usage{
comp.cat.counts(long.df, var, netID = "netID", fun = fun.count)
}
\arguments{
\item{long.df}{A 'long' dataframe with alteri/dyads in rows.}

\item{var}{Alter attribute which categories are to be counted.}

\item{netID}{Name of the network-/ ego-ID variable in \code{broad}.}

\item{fun}{Function to be used for counting. \code{fun.count} for absolute counts \code{fun.prop} propotional counts.}
}
\value{
returns a \code{dataframe} with counts of all categories as variables.
}
\description{
This function counts the category frequencies (absolute or proportional) of a variable representing alter attributes in ego-centric-network data.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

