% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{comp.diversity}
\alias{comp.diversity}
\title{Calculate the network diversity of ego-centered networks for a given alter-attribute.}
\usage{
comp.diversity(cat.counts, netsize)
}
\arguments{
\item{cat.counts}{A category-count-dataframe usually generated with comp.cat.counts().}

\item{netsize}{A vector of network sizes.}
}
\value{
Returns the absolute and proportional count of unique categories per ego/ network in a \code{dataframe}.
}
\description{
This function calculates the network diversity of ego-centered networks for a given alter-attribute.
}
\keyword{internal}
