% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composition.R
\name{composition}
\alias{composition}
\title{Calculate compositional measurements (Proportion, Diversity, EI-Index)}
\usage{
composition(alters, v_alt, netsize, egoID = "egoID", v_ego = NULL,
  mode = "regular")
}
\arguments{
\item{alters}{\code{Data frame} containing alters data (alters as cases),
alters are separated by a variable containg an egoID.}

\item{v_alt}{A character naming the variable containg the alter-attribute.}

\item{netsize}{\code{Vector} containing values of network size per ego.}

\item{egoID}{\code{Character}. Name of the variable identifying egos (default: "egoID").}

\item{v_ego}{Character vector containing the ego attribute. Only needed for homophily index (EI). Caution: Levels of v_alt and v_ego need to correspond (see Details).}

\item{mode}{A character. "regular" for a basic output, "all" for a complete output.}
}
\value{
Returns a dataframe with category counts, diversity and EI-Index values in an ego-centered network for a provided alter attribute.
}
\description{
This function outputs a dataframe containg serveral compositional measures
for ego-centered-network data. It reports the proportion of each group in the
network, the absolute count of groups present and, if provided the
corresponding ego attribute, the EI-Index is employed as a measurment for ego's tendency
to homo-/heterophily.
}
\details{
v_ego is expected to consist of one entry per ego. The ego
attributes are usually drawn from the ego dataframe. If the ego attribute is
stored alongside the alters data, make sure to drop repeated values per
alters (see example two).
}
\examples{
# Load example data
data("egos32")
data("alters32")

# Example one
composition(alters32, v_alt = "alter.sex", netsize = egos32$netsize, 
            v_ego = egos32$sex)

# Example two
# - using an ego attribute stored in the alter dataframe.
ego.sex <- alters32[!duplicated(alters32$egoID), ]$ego.sex
res <- composition(alters32, v_alt = "alter.sex", netsize = egos32$netsize, 
                   v_ego = ego.sex)

# Using cbind() to show the ego attribute alongside the results might
# be helpful in many cases:
cbind(res, ego.sex)
}
\keyword{analysis}
\keyword{ego-centered}
\keyword{network}
