% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_network_igrpah.R
\name{convert_egor}
\alias{convert_egor}
\alias{to_network}
\alias{to_igraph}
\title{Convert egor object to network or igraph objects}
\usage{
to_network(object, directed = F, include.ego = F, ego.attrs = NULL,
  ego.alter.weights = NULL)

to_igraph(object, directed = F, include.ego = F, ego.attrs = NULL,
  ego.alter.weights = NULL)
}
\arguments{
\item{object}{An egor object.}

\item{include.ego}{\code{Logical.} Should ego be included?}

\item{ego.attr}{Vector of names (character) or indices (numeric) of ego
variables that should be carried over to  the network/
igraph objects.}

\item{ego.alter.attr}{Vector of names (character) or indices (numeric) of
alter variables that should be carried over to the the
network/ igraph objects, as edge attributes of the ego-alter relations.}
}
\description{
These functions turns an egor object into a list of network or igraph objects.
By default ego itself is not included in the created objects, but
there is a parameter that allow for including ego. By specifying variable
names on the  ego level, that correspond to alter variables and
attributes of ego-alter relations, matching ego and alter variables can be
included (see details).
}
\details{
The names of the variables specified in ego.attr and ego.alter.attr
need to be the same as the names of corresponding alter attributes,
in order for those variables to be complete in the resulting network/ igraph
object (see example).
}
\section{Functions}{
\itemize{
\item \code{to_network}: Creates a lis of statnets network objects.

\item \code{to_igraph}: Creates a list of igraph objects.
}}

