% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dyad.poss}
\alias{dyad.poss}
\title{Calculate possible dyads for a given number of alteri.}
\usage{
dyad.poss(max.alteri)
}
\arguments{
\item{alter.folder}{}

\item{edge.folder}{}
}
\description{
Here should be a a longer description of this function.
}
\keyword{ego-centric}
\keyword{network}
\keyword{sna}

