% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.random.egonet.data.R
\name{edgelist_to_wide}
\alias{edgelist_to_wide}
\title{Transforms edge lists to alter-alter wide format data.}
\usage{
edgelist_to_wide(edges)
}
\arguments{
\item{edges}{List of \code{data.frames} containing edge lists.}
}
\description{
Only works properly, if the netsize of all networks is constant.
}
\keyword{ego-centered}
\keyword{internal}
\keyword{network}
