% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{edges.attributes.to.network}
\alias{edges.attributes.to.network}
\title{edges.attributes.to.network}
\usage{
edges.attributes.to.network(e.list, alteri)
}
\arguments{
\item{e.list}{\code{data.frame} containg edge data/ one edgelist.}

\item{alteri}{\code{data.frame} containg alteri attributes.}
}
\description{
This function generates one igraph object from an edgelist and a dataframe 
alteri attributes.
}
\keyword{internal}

