% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{edges.attributes.to.network}
\alias{edges.attributes.to.network}
\title{edges.attributes.to.network}
\usage{
edges.attributes.to.network(e.list, alters)
}
\arguments{
\item{e.list}{\code{data.frame} containg edge data/ one edgelist.}

\item{alters}{\code{data.frame} containg alter attributes.}
}
\description{
This function generates one igraph object from an edgelist and a dataframe
containing alter attributes.
}
\keyword{internal}
