% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egor_doc.R
\docType{package}
\name{egor-package-doc}
\alias{egor-package-doc}
\title{\code{egor}}
\description{
R Package for importing and analyzing ego-centered-network data.
}
\details{
\href{https:///egor.tillt.net/}{Further Information}
or \href{https://github.com/tilltnet/egor}{GitHub}

When analyzing ego-centered network data it is common to either
include or exclude data on ego. By default \code{egor} excludes ego for most
analytic and visual functions and offers to include ego with the argument
\code{include.ego}. In order for egor to automatically detect which ego-level variables
correspond to the alter-variables, the variables on the different levels need
to be named exactly the same. Alternatively the functions that have an
\code{include.ego} argument have additional arguments that allow to specify the
name of the ego-level variable in question.
}
\author{
Till Krenz, \email{egor@tillt.net}

Pavel Krivitsky, \email{pavel@uow.edu.au}

Michał Bojanowski \email{m.bojanowski at icm.edu.pl}

Andreas Herz, \email{herzand@uni-hildesheim.de}

Raffaele Vacca, \email{r.vacca@ufl.edu}
}
