% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egor_doc.R
\docType{package}
\name{egor-package-doc}
\alias{egor-package-doc}
\title{\code{egor}}
\description{
R Package for importing and analyzing ego-centered-network data.
}
\details{
\href{https://tilltnet.github.io/egor/}{Further Information}
or \href{https://github.com/tilltnet/egor}{GitHub}

Thanks to: Martina Morris
}
\author{
Till Krenz, \email{egor@tillt.net}

Pavel Krivitsky, \email{pavel@uow.edu.au}

Michał Bojanowski \email{m.bojanowski at icm.edu.pl}

Andreas Herz, \email{herzand@uni-hildesheim.de}

Raffaele Vacca, \email{r.vacca@ufl.edu}

Christopher McCarty, \email{ufchris@ufl.edu}

Markus Gamper, \email{m.gamper@uni-koeln.de}
}
