% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{egoR.density}
\alias{egoR.density}
\title{Calculate the density for n+many ego-centered networks.}
\usage{
egoR.density(egoR, weight = NULL, max.netsize = NULL, directed = FALSE)
}
\arguments{
\item{egoR}{A list of six data objects, created by one of egonetR's import 
functions.}

\item{max.netsize}{Optional parameter. Only needed if the density information
was only collected for a subsample of all alteri per network.}

\item{directed}{logical indicating if the alter-alter relation data/ edges 
are directed or undirected.}
}
\value{
returns a \code{vector} of network density values.
}
\description{
This function uses an 'egoR' object (list of data objects created by an 
egonetR import function) and calculates the density of all 
the ego-centered networks listed in the 'egoR' object.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

