% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egor.R
\name{egor}
\alias{egor}
\alias{as.egor}
\title{egor - a data class for ego-centered network data.}
\usage{
egor(alters.df, egos.df = NULL, alter_ties.df = NULL, egoID = "egoID",
  ego.design = list(~1), alter.design = list(max = Inf))

as.egor(x, ...)
}
\arguments{
\item{alters.df}{either a \code{data.frame} containing the alters
(whose nominator is identified by the column specified by \code{egoID}
or a list of data frames with the same columns, one for each ego,
with empty data frames or \code{NULL}s corresponding to egos with no
nominees.}

\item{egos.df}{\code{data.frame} containing the egos.}

\item{alter_ties.df}{\code{data.frame} containing the alter-alter
relations in the style of an edge list, or a list of data frames
similar to \code{alters.df}.}

\item{egoID}{Name of ego ID variable; optional if \code{alters.df} and
\code{alter_ties.df} are both lists of data frames.}

\item{ego.design}{A \code{\link{list}} of arguments to
\code{\link{svydesign}} specifying the sampling design for the
egos. If formulas, they can refer to columns of \code{egos.df}.}

\item{alter.design}{A \code{\link{list}} of arguments specifying
nomination information. Currently, the following elements are
supported: \describe{\item{\code{"max"}}{Maximum number of alters
that an ego can nominate.}}}

\item{x}{an object to be coerced to \code{\link{egor}}.}
}
\value{
An \code{\link{egor}} object. An \code{\link{egor}} is a \code{\link{tibble}} whose
top-level columns store the ego attributes, and which has two
special nested columns: \code{.alters}, containing, for each row (ego)
a table of that ego's alter attributes and \code{.alter_ties}, a table
containing that ego's alter--alter ties, if observed.

In addition, it has two attributes: \code{ego.design}, containing an
object returned by \code{\link[survey:svydesign]{survey::svydesign()}} specifying the sampling
design by which the egos were selected and \code{alter.design}, a
\code{\link{list}} containing specification of how the alters were
nominated. See the argument above for currently implemented
settings.
}
\description{
The function \code{\link[=egor]{egor()}} is used to create an egor object from
ego-centered network data.
}
\details{
If parameters \code{alters.df}, \code{egos.df}, and \code{alter_ties.df}
are data frames, they need to share a common ego ID variable,
with corresponding values. If \code{alters.df} and \code{alter_ties.df} are
lists of data frames, \code{egoID} is ignored and they are matched
positionally with the rows of \code{egos.df}. Of the three parameters
only \code{alters.df} is necessary to create an \code{egor} object, and
\code{egos.df} and \code{alter_ties.df} are optional.
}
\examples{

}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}
