% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizz.R
\name{egoR.vis.wizzard}
\alias{egoR.vis.wizzard}
\title{Network Visualisation Wizzard (egoR)}
\usage{
egoR.vis.wizzard(egoR)
}
\arguments{
\item{egoR}{A list of six data objects, created by one of egor's import 
functions.}
}
\value{
Opens an interactive Browser Application.
}
\description{
This function uses an 'egoR' object (list of data objects created by an 
egor import function), extracts the graphs object and uses it to visualise
all networks contained in an interactive Browser Application (R-Shiny).
}
\examples{
data("egoR32")
egoR.vis.wizzard(egoR32)
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}
