% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egos32}
\alias{egos32}
\title{32 randomly created egos belonging to ego-centered networks}
\format{A data frame with 32 sets of alter-alter relations and 4 variables:
\describe{
\item{.EGOID}{ego identifier}
\item{age}{age in categories}
\item{age.years}{age in years}
\item{country}{country}
\item{income}{income}
\item{sex}{gender}
}}
\usage{
egos32
}
\description{
32 randomly created egos belonging to ego-centered networks
}
\keyword{datasets}
