% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ei.R
\name{EI}
\alias{EI}
\title{Calculate the EI-Index}
\usage{
EI(alteri, edges, var_name, egoID = "egoID", alterID = "alterID")
}
\arguments{
\item{alteri}{}

\item{edges}{\code{List} of edgelist-\code{dataframes} or one
\code{dataframes} #' containing all edges,}
}
\description{
The EI-Index compares the inner-group edge density to the outer-group edge
density. It can be calculated for the whole network and for groups. If you
want to calculate ego's EI use the composition command.
}
\keyword{ego-centric}
\keyword{network}
\keyword{sna}

