% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei.R
\name{EI}
\alias{EI}
\title{Calculate the EI-Index}
\usage{
EI(object, var_name)
}
\arguments{
\item{object}{An \code{egor} object.}

\item{var_name}{\code{Character} naming grouping variable.}

\item{aaties}{\code{data.frame/ list} containg alter-alter relations as a 
'global edge list' or as a list of 'edge lists'.
(not needed if \code{object} is an \code{egor} object).}
}
\description{
The EI-Index is the division of the intra-group edge density and the outer-group edge
density. It can be calculated for the whole network and for subgroups. The
whole network EI is a metric indicating the tendency of a network to be
clustered by the categories of a given factor variable. The EI value of a
groups describes the tendency of a group to be connected or not connected
to other groups. Additionally, the EI index can be employed as a measurment
for egos tendendy to homo-/heterphily - use the \code{composition} command
for individual EI-Index.
}
\examples{
data("alters32")
data("edges32")
EI(alters32, edges32, var_name = "alter.sex", altID = "alterID")
}
\references{
Krackhardt, D., Stern, R.N., 1988. Informal networks and
organizational crises: an experimental simulation. Social Psychology
Quarterly 51 (2), 123-140.

Everett, M. G., & Borgatti, S. P. (2012). Categorical attribute
based centrality: E-I and G-F centrality. Social Networks, 34(4), 562-569.
}
\keyword{ego-centered}
\keyword{network}
\keyword{sna}
