% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ei.R
\name{fragmentations}
\alias{fragmentations}
\title{Network fragments of ego-centerd networks}
\usage{
fragmentations(alters.list, edges.list)
}
\arguments{
\item{alters.list}{\code{List} of \code{data frames} containing the alters 
data.}

\item{edges.list}{\code{List} of \code{data frames} containing the edge 
lists (= alter-alter relations).}
}
\description{
Calculate the count of fragments ego-centered networks form, if their 
respective egos are removed from the network.
}
\keyword{ego-centered}
\keyword{network}
\keyword{sna}
