% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ei.R
\name{fragmentations}
\alias{fragmentations}
\title{Fragmentations of a list of ego-centered networks}
\usage{
fragmentations(edges_)
}
\arguments{
\item{edges_}{\code{List} of edgelist-\code{dataframes}}
}
\description{
This gives the number of seperate fragments in a \code{list} ego-centered
networks.
}
\keyword{ego-centered}
\keyword{network}
\keyword{sna}

