% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/composition.R
\name{fun.count}
\alias{fun.count}
\title{Generate counting variables for each category of a variable. Used by comp.cat.counts().}
\usage{
fun.count(x)
}
\arguments{
\item{x}{A factor representing an alteri-attribute variable.}
}
\description{
This function generates Variables counting the frequency of each category of a variable. To use on data aggregated from the long ego-centric network data format.
#'
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

