% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.random.egonet.data.R
\name{generate.sample.edge.list}
\alias{generate.sample.edge.list}
\title{Generate a random edge list for one network.}
\usage{
generate.sample.edge.list(netsize)
}
\arguments{
\item{netsize}{\code{Numeric} indicating maximum number of alters.}
}
\description{
Generate a random edge list for one network.
}
\keyword{ego-centric}
\keyword{internal}
\keyword{network}
