% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.sample.edge.list}
\alias{generate.sample.edge.list}
\title{Generate a random edge list for one network.}
\usage{
generate.sample.edge.list(max.alteri)
}
\arguments{
\item{alter.folder}{}

\item{edge.folder}{}
}
\description{
Here should be a a longer description of this function.
}
\keyword{ego-centric}
\keyword{network}
\keyword{sna}

