% Generated by roxygen2 (4.0.2): do not edit by hand
\name{generate.sample.ego.data}
\alias{generate.sample.ego.data}
\title{Generate random ego-centric-network data.}
\usage{
generate.sample.ego.data(net.count, max.alteri)
}
\arguments{
\item{net.count}{Number of networks/ egos to generate.}

\item{max.alteri}{Maximum size of networks.}
}
\description{
This funtion generates random ego-centric-network data for a specified number of networks with a maximum network size.
}
\keyword{ego-centric}
\keyword{network}
\keyword{sna}

