% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{long.df.to.list}
\alias{long.df.to.list}
\title{Trim/ listify ego-centric network data}
\usage{
long.df.to.list(long, wide, netsize, egoID, back.to.df = F)
}
\arguments{
\item{long}{A 'long' dataframe with alteri/dyads in rows.}

\item{wide}{A \code{data.frame} of alteri attributes in the wide format.}

\item{netsize}{\code{vector} containing values of network size per ego.}

\item{egoID}{Name of the variable identifying egos (default: "egoID").}

\item{back.to.df}{If \code{TRUE} a dataframe is returned, if \code{FALSE} a
list. Defaults to \code{FALSE}.}
}
\value{
Returns a \code{list} of \code{dataframes} where every
\code{dataframe} represents one ego/ network and the rows in the
\code{dataframe} represents one alter. If the \code{back.to.df} parameter is
called the \code{list} entries are combined to one \code{dataframe}, in the
'long' format.
}
\description{
This function generates the \code{alteri.list} object. \code{alteri.list} is a list where
each entry entails a \code{dataframe} of the alteri of one ego. By using
the \code{netsize} variable it is ensured, that the list entries are of the
correct length and possibly present rows of NA values are deleted.
}

