% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.egonet.folders}
\alias{read.egonet.folders}
\title{Import ego-centric network data from separate folders for edgelist and
alteri-attributes.}
\usage{
read.egonet.folders(egos.file, alter.folder, edge.folder, netID = "netID")
}
\arguments{
\item{alter.folder}{}

\item{edge.folder}{}

\item{netID}{}

\item{egos}{}
}
\description{
This function imports ego-centric network data from folders with separate
files for alteri-level data and dyads.
}
\keyword{ego-centric}
\keyword{network,}
\keyword{sna}

