% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.ego.folders.R
\name{read.egonet.folders}
\alias{read.egonet.folders}
\title{Import ego-centric network data from separate folders for edgelist and 
alteri-attributes.}
\usage{
read.egonet.folders(egos.file, alter.folder, edge.folder, csv.sep = ",",
  egoID = "egoID", first.col.row.names = FALSE)
}
\arguments{
\item{egos.file}{File name of the .csv file containg the ego data.}

\item{alter.folder}{Folder name of the folder containing the alter data in
separate .csv files for each ego/ network.}

\item{edge.folder}{Folder name of the folder containing the edge/ tie data in
separate .csv files for each ego/ network.}

\item{csv.sep}{\code{Character} indicating the seperator used in csv files.}

\item{egoID}{\code{Character}. Name of the variable identifying egos (default: "egoID").}

\item{first.col.row.names}{\code{Boolean} indicating if first column contains
row names, that are to be skipped, default is \code{FALSE}.}
}
\value{
A \code{list} of six objects - the **egoR** object: (1) egos.df: \code{dataframe} of all
egos and their attributes; (2) alteri.df: \code{dataframe} of all alteri;
(3) alteri.list: \code{list} of \code{dataframes} of all alteri per ego;
(4) edges: \code{list} of \code{dataframes} with edge lists per network; (5)
graphs: \code{list} of igraph objects; (6) results: a result \code{dataframe},
pre-populated with the network size of each network
}
\description{
This function imports ego-centric network data from folders with separate 
files for alteri-level and edge data. It will run some basic checks upon
the completness of the data and inform the user of potential problems. This
functions can be used to import data exported from EgoNet (McCarty 2011).
}
\examples{
egos.file <-  system.file("extdata", "egos_32.csv", package = "egor")
alteri.folder <- system.file("extdata", "alteri_32", package = "egor")
edge.folder <-  system.file("extdata", "edges_32", package = "egor")

ef <- read.egonet.folders(egos.file = egos.file, alter.folder = alteri.folder, edge.folder = edge.folder, csv.sep = ";")
}
\keyword{ego-centric}
\keyword{import}

