% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{read.egonet.one.file}
\alias{read.egonet.one.file}
\title{Import ego-centric network data from 'one file format'}
\usage{
read.egonet.one.file(egos, netsize, egoID = "egoID", attr.start.col,
  attr.end.col, dy.max.alters, dy.first.var, ego.vars = NULL, var.wise = F,
  ...)
}
\arguments{
\item{egos}{\code{Data frame} containg ego data (egos as cases)}

\item{netsize}{\code{Vector} containing values of network size per ego.}

\item{egoID}{\code{Character}. Name of the variable identifying egos (default: "egoID").}

\item{attr.start.col}{First colum containing alter attributes.}

\item{attr.end.col}{Last colum containing alter attributes.}

\item{dy.max.alters}{Maximum number of alters.}

\item{dy.first.var}{First column containing alter-alter relations/ edges.}

\item{ego.vars}{A \code{data.frame} of alter attributes in the wide format.}

\item{var.wise}{Logical value indicatin if the alter attributes are sorted variable wise (defaults to FALSE).}

\item{...}{additional arguments to \code{\link[=egor]{egor()}}.}
}
\value{
A \code{list} of six objects - the **egoR** object: (1) egos.df: \code{dataframe} of all
egos and their attributes; (2) alters.df: \code{dataframe} of all alters;
(3) alters.list: \code{list} of \code{dataframes} of all alters per ego;
(4) edges: \code{list} of \code{dataframes} with edge lists per network; (5)
graphs: \code{list} of igraph objects; (6) results: a result \code{dataframe},
pre-populated with the network size of each network
}
\description{
This function imports ego-centric network data, stored in a single file, providing
ego, alter and edge data. This data format is for exampled used by the Allbus 2010 (GESIS)
and similar social surveys.
}
\references{
Muller, C., Wellman, B., & Marin, A. (1999). How to Use SPSS to
Study Ego-Centered Networks. Bulletin de Methodologie Sociologique,
64(1), 83-100.
}
\keyword{import}
