% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read.egonet.one.file}
\alias{read.egonet.one.file}
\title{Import ego-centric network data from 'one file format'}
\usage{
read.egonet.one.file(egos, netsize, egoID = "egoID", attr.start.col,
  attr.end.col, dy.max.alteri, dy.first.var)
}
\arguments{
\item{egos}{\code{dataframe} containg ego data (egos as cases)}

\item{netsize}{\code{vector} containing values of network size per ego.}

\item{egoID}{Name of the variable identifying egos (default: "egoID").}

\item{attr.start.col}{}

\item{dy.max.alteri}{}

\item{dy.first.var}{}
}
\value{
A \code{list} of six objects: (1) egos.df: \code{dataframe} of all
egos and their attributes; (2) alteri.df: \code{dataframe} of all alteri;
(3) alteri.list: \code{list} of \code{dataframes} of all alteri per ego;
(4) edges: \code{list} of \code{dataframes} with edge lists per network; (5)
graphs: \code{list} of igraph objects; (6) results: a result \code{dataframe},
pre-populated with the network size of each network
}
\description{
This function imports ego-centric network data, from one file, providing
ego, alter and edge data.
}

