% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{read.egonet.two.files}
\alias{read.egonet.two.files}
\title{Import ego-centric network data from two file format}
\usage{
read.egonet.two.files(egos, alteri, netsize = NULL, egoID = "egoID",
  alterID = NULL, e.max.alteri, e.first.var, ego.vars = NULL,
  selection = NULL)
}
\arguments{
\item{egos}{\code{dataframe} containg ego data (egos as cases)}

\item{alteri}{\code{dataframe} containing alteri data (alteri as cases),
alteri are separated by a variable containg an egoID.}

\item{netsize}{\code{vector} containing values of network size per ego.}

\item{egoID}{\code{Character}. Name of the variable identifying egos (default: "egoID").}

\item{alterID}{\code{Character}. Name of the variable identifying egos (default: "alterID").}

\item{e.max.alteri}{Maximum number of alteri that are included in edge data.}

\item{e.first.var}{Index of first column in \code{egos} containing edge data.}

\item{ego.vars}{\code{Character vector} naming variables in the egos data,
in order to copy them in to the long alteri \code{dataframe}.}

\item{selection}{\code{Character} naming \code{numeric} variable indicating 
alteri selection with zeros and ones.}
}
\value{
A \code{list} of six objects - the **egoR** object: (1) egos.df: \code{dataframe} of all
egos and their attributes; (2) alteri.df: \code{dataframe} of all alteri;
(3) alteri.list: \code{list} of \code{dataframes} of all alteri per ego;
(4) edges: \code{list} of \code{dataframes} with edge lists per network; (5)
graphs: \code{list} of igraph objects; (6) results: a result \code{dataframe},
pre-populated with the network size of each network
}
\description{
This function imports ego-centric network data, stored in two files, where 
one file contains the ego attributes and the edge information and the other file 
contains the alteri data. This form of data storage for ego-centered network data 
is proposed by Muller, Wellman and Marin (1999).
}
\keyword{import}

