% Generated by roxygen2 (4.0.2): do not edit by hand
\name{to.network}
\alias{to.network}
\title{to.network}
\usage{
to.network(elist, attributes)
}
\arguments{
\item{elist}{[list of dataframes]}

\item{attributes}{[list of dataframes]}
}
\description{
This function generates a list of igraph objects from an edgelist and a
dataframe alteri attributes.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{netowrk}

