% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.egonet.R
\name{to.network}
\alias{to.network}
\title{to.network}
\usage{
to.network(e.lists, alteri.list)
}
\arguments{
\item{e.lists}{\code{List} of \code{data.frame}s containg edge data/ one edgelist.}

\item{alteri.list}{\code{List} of \code{data.frame}s containg alteri attributes.}
}
\description{
This function generates a list of igraph objects from an edgelist and a
dataframe alteri attributes.
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

