% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_methods.r
\name{trim_alters}
\alias{trim_alters}
\title{Trims alters that are missing/ deleted from ego data}
\usage{
trim_alters(object)
}
\arguments{
\item{object}{An \code{egor} object.}
}
\value{
An \code{egor} object with trimmed alter-alter ties (.aaties).
}
\description{
Trims alters that are missing/ deleted from ego data
}
