% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wizz.R
\name{vis.wizzard}
\alias{vis.wizzard}
\title{Network Visualisation Wizzard}
\usage{
vis.wizzard()
}
\arguments{
\item{egoR}{A list of six data objects, created by one of egonetR's import 
functions.}

\item{max.netsize}{Optional parameter. Only needed if the density information
was only collected for a subsample of all alteri per network.}
}
\value{
returns a \code{vector} of network density values.
}
\description{
This function uses an object named 'graphs' (list of data objects created by an 
egonetR import function), extracts the graphs object and uses it to visualise
all networks contained (Shiny App).
}
\keyword{analysis}
\keyword{ego-centric}
\keyword{network}

