% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego.design.R
\name{weights.egor}
\alias{weights.egor}
\title{\code{\link[=weights.egor]{weights.egor()}} extracts the (relative) sampling weights of each
ego in the dataset.}
\usage{
\method{weights}{egor}(object, ...)
}
\arguments{
\item{object}{an \code{\link{egor}} object.}

\item{...}{arguments to be passed to methods}
}
\description{
\code{\link[=weights.egor]{weights.egor()}} extracts the (relative) sampling weights of each
ego in the dataset.
}
\seealso{
\code{\link[survey]{weights.survey.design}}
}
